+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning]https://semver.org/spec/v2.0.0.
 
 ### Unreleased
 
+- fix: processed block wait load cache address from db for first start
+
 ### [0.9.9] - 2026-01-23
 
 - fix: increase default page size for transaction