+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning]https://semver.org/spec/v2.0.0.
 
 ### Unreleased
 
+- feat: redesign README.md [DV-4185]
 - fix: processed block wait load cache address from db for first start
 
 ### [0.9.9] - 2026-01-23