+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning]https://semver.org/spec/v2.0.0.
 ## Releases
 
 ### Unreleased
+- Added rollback support [DV-1943]
 
 - feat: redesign README.md [DV-4185]
 - fix: processed block wait load cache address from db for first start